Move ExpressionArray to Lists
